requirements.txt
# Django Framework
Django==4.2.7

# Database
mysqlclient==2.2.0
dj-database-url==2.1.0

# Django REST Framework
djangorestframework==3.14.0

# Environment variables
python-decouple==3.8

# Image processing
Pillow==10.0.1

# Static files for production
whitenoise==6.6.0

# HTTP requests for external API
requests==2.31.0

# Production server
gunicorn==21.2.0

# psycopg2 for PostgreSQL (Render uses PostgreSQL)
psycopg2-binary==2.9.9
